# implementations/go/examples/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the DDD-based Account domain example service (Account/Card/Payment/Refund/Credential Bounded Contexts). Generated from the @-annotations above each handler in internal/interface/http/ by swag (docs/architecture/api-documentation.md) — never hand-edited.",
        "title": "Account Service API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/auth/sign-in": {
            "post": {
                "description": "Verifies the user ID/password and issues a JWT bearer token. Responds with the same error whether the user ID doesn't exist or the password is wrong, to prevent user enumeration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in",
                "parameters": [
                    {
                        "description": "User ID and password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.SignInRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "A bearer token was issued.",
                        "schema": {
                            "$ref": "#/definitions/http.SignInResponse"
                        }
                    },
                    "400": {
                        "description": "Request validation failed (`VALIDATION_FAILED`) — e.g. a missing `userId`/`password`.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The user ID doesn't exist, or the password doesn't match (`INVALID_CREDENTIALS`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sign-up": {
            "post": {
                "description": "Registers a new user ID with a password. The password is hashed before storage; the plaintext is never persisted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Sign up",
                "parameters": [
                    {
                        "description": "New user ID and password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The user was registered."
                    },
                    "400": {
                        "description": "One of: request validation failed (`VALIDATION_FAILED`) — e.g. a missing `userId` or a password under 8 characters — or the user ID is already taken (`USER_ID_ALREADY_EXISTS`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Issues a new card linked to an active account owned by the authenticated requester.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Card"
                ],
                "summary": "Issue a card",
                "parameters": [
                    {
                        "description": "Linked account and card brand",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.IssueCardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "The card was issued.",
                        "schema": {
                            "$ref": "#/definitions/http.CardResponse"
                        }
                    },
                    "400": {
                        "description": "One of: request validation failed (`VALIDATION_FAILED`), or the linked account is not active (`CARD_ISSUE_REQUIRES_ACTIVE_ACCOUNT`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No account exists with the given `accountId` for this requester (`LINKED_ACCOUNT_NOT_FOUND`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/{cardId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the card only if it belongs to the authenticated requester.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Card"
                ],
                "summary": "Look up a card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The card ID",
                        "name": "cardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The card was found.",
                        "schema": {
                            "$ref": "#/definitions/http.CardResponse"
                        }
                    },
                    "401": {
                        "description": "The bearer token is missing, malformed, or invalid.",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No card exists with the given `cardId` for this requester (`CARD_NOT_FOUND`).",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.CardResponse": {
            "type": "object",
            "properties": {
                "accountId": {
                    "description": "The account this card is linked to.",
                    "type": "string"
                },
                "brand": {
                    "description": "The card brand (e.g. `VISA`, `MASTER`).",
                    "type": "string"
                },
                "cardId": {
                    "description": "The card's ID.",
                    "type": "string"
                },
                "createdAt": {
                    "description": "When the card was issued.",
                    "type": "string"
                },
                "ownerId": {
                    "description": "The card owner's user ID.",
                    "type": "string"
                },
                "status": {
                    "description": "The card status (`ACTIVE`, `SUSPENDED`, or `CANCELLED`).",
                    "type": "string"
                }
            }
        },
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "A stable, machine-readable error code the client can branch on (e.g. `ACCOUNT_NOT_FOUND`). Unlike `message`, this never changes wording.",
                    "type": "string"
                },
                "error": {
                    "description": "The standard HTTP status text for `statusCode` (e.g. `Not Found`).",
                    "type": "string"
                },
                "message": {
                    "description": "A human-readable description of the error.",
                    "type": "string"
                },
                "statusCode": {
                    "description": "The HTTP status code.",
                    "type": "integer"
                }
            }
        },
...
        "http.IssueCardRequest": {
            "type": "object",
            "properties": {
                "accountId": {
                    "description": "The active account to link the card to. Must be owned by the authenticated requester.",
                    "type": "string"
                },
                "brand": {
                    "description": "The card brand (e.g. `VISA`, `MASTER`).",
                    "type": "string"
                }
            }
        },
...
        "http.SignInRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "The account password.",
                    "type": "string"
                },
                "userId": {
                    "description": "The registered user ID.",
                    "type": "string"
                }
            }
        },
        "http.SignInResponse": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "description": "A JWT bearer token. Send it as `Authorization: Bearer \u003caccessToken\u003e` on subsequent requests.",
                    "type": "string"
                }
            }
        },
        "http.SignUpRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "description": "The account password. Must be at least 8 characters.",
                    "type": "string"
                },
                "userId": {
                    "description": "The user ID to register. Must be unique.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token issued by `POST /auth/sign-in`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}